edit backend flow add jwt token

+++ plutocart_backend/src/main/java/com/example/plutocart/services/AccountService.java
@@ -1,11 +1,13 @@
 package com.example.plutocart.services;
 
+import com.example.plutocart.auth.JwtUtil;
 import com.example.plutocart.dtos.account.AccountDTO;
 import com.example.plutocart.entities.Account;
 import com.example.plutocart.repositories.AccountRepository;
 import com.example.plutocart.security.IMEIEncryption;
 import com.example.plutocart.utils.GenericResponse;
 import com.example.plutocart.constants.ResultCode;
+import io.jsonwebtoken.Jwts;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -31,6 +33,9 @@ public class AccountService {
            accountDTO.setAccountId(account.getAccountId());
            response.setStatus(ResultCode.SUCCESS);
            response.setData(accountDTO);
+           String acId = String.valueOf(accountDTO.getAccountId());
+           String token = JwtUtil.generateToken(acId);
+           response.setAuthentication(token);
            return response;
        }
        catch (Exception ex){
@@ -53,6 +58,9 @@ public class AccountService {
                 accountDTO.setAccountId(account.getAccountId());
                 response.setStatus(ResultCode.SUCCESS);
                 response.setData(accountDTO);
+                String acId = String.valueOf(accountDTO.getAccountId());
+                String token = JwtUtil.generateToken(acId);
+                response.setAuthentication(token);
                 return response;
             }
 
@@ -63,35 +71,53 @@ public class AccountService {
         return response;
     }
 
-    public GenericResponse UpdateAccountToMember(String email , Integer accountId){
+    public GenericResponse UpdateAccountToMember(String email , Integer accountId , String token){
         GenericResponse response = new GenericResponse();
-        try {
-            accountRepository.updateAccountToMember(email , accountId);
-            response.setStatus(ResultCode.SUCCESS);
-            Account account = accountRepository.findById(accountId).get();
-            account.setAccountRole("Member");
-            account.setEmail(email);
-            response.setData(account);
-            return response;
+        String userId = JwtUtil.extractUsername(token);
+        if(Integer.parseInt(userId) == accountId){
+            try {
+                accountRepository.updateAccountToMember(email , accountId);
+                response.setStatus(ResultCode.SUCCESS);
+                Account account = accountRepository.findById(accountId).get();
+                account.setAccountRole("Member");
+                account.setEmail(email);
+                response.setData(account);
+                return response;
+            }
+            catch (Exception ex){
+                response.setStatus(ResultCode.BAD_REQUEST);
+                response.setData(null);
+                return response;
+            }
         }
-        catch (Exception ex){
-            response.setStatus(ResultCode.BAD_REQUEST);
+        else {
+            response.setStatus(ResultCode.FORBIDDEN);
             response.setData(null);
             return response;
         }
+
     }
 
-    public GenericResponse DeleteAccount(Integer accountId){
+    public GenericResponse DeleteAccount(Integer accountId , String token){
         GenericResponse response = new GenericResponse();
-        try {
-            accountRepository.deleteAccount(accountId);
-            response.setStatus(ResultCode.SUCCESS);
-            return response;
+        String userId = JwtUtil.extractUsername(token);
+        System.out.println("using DeleteAccount Service");
+        if(Integer.parseInt(userId) == accountId){
+            try {
+                accountRepository.deleteAccount(accountId);
+                response.setStatus(ResultCode.SUCCESS);
+                return response;
+            }
+            catch (Exception ex){
+                response.setStatus(ResultCode.BAD_REQUEST);
+                response.setData(null);
+                return response;
+            }
         }
-        catch (Exception ex){
-            response.setStatus(ResultCode.BAD_REQUEST);
-            response.setData(null);
-            return response;
+       else {
+           response.setStatus(ResultCode.FORBIDDEN);
+           response.setData(null);
+           return  response;
         }
     }
